# src/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Esta es la documentación para uso del API para la aplicación ramspace",
    "version": "1.0.0",
    "title": "RamSpace API",
    "contact": {
      "email": "jdany.mc@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "schemes": [
    "http"
  ],
  "host": "localhost:3000",
  "basePath": "/api",
...
  "paths": {
...
    "/articulo": {
      "get": {
        "tags": [
          "Articulo"
        ],
        "summary": "Obtiene todos los Articulos",
        "description": "Obtiene todos los Articulos registrados",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "content": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/ArticuloResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Articulo"
        ],
        "summary": "Guarda un Articulo",
        "description": "Guarda un Articulo",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Objeto Articulo",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Articulo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "content": {
                  "$ref": "#/definitions/ArticuloResponse"
                }
              }
            }
          },
          "404": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/articulo/{id}": {
      "get": {
        "tags": [
          "Articulo"
        ],
        "summary": "Obtiene datos de un Articulo",
        "description": "Obtiene datos de un Articulo",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id de Articulo para mostrar",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "content": {
                  "$ref": "#/definitions/ArticuloIdResponse"
                }
              }
            }
          },
          "404": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Articulo"
        ],
        "summary": "Actualiza datos de un Articulo",
        "description": "Actualiza datos de un Articulo",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id de Articulo para actualizar",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Objeto Articulo",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Articulo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "content": {
                  "$ref": "#/definitions/ArticuloResponse"
                }
              }
            }
          },
          "404": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Articulo"
        ],
        "summary": "Elimina un Articulo",
        "description": "Elimina un Articulo",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id de Articulo a eliminar",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "type": "string",
              "example": "OK"
            }
          },
          "404": {
            "description": "Invalid Status Value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/articulo/busqueda": {
      "post": {
        "tags": [
          "Articulo"
        ],
        "summary": "Obtiene artículos segun un parametro de busqueda",
        "description": "Obtiene articulos segun un parametro de busqueda",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Objeto parametro de busqueda",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "busqueda": {
                  "type": "string",
                  "example": "monitor"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "content": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Articulo_ImagenResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    },
    "/articulo/categoria/{id}": {
      "get": {
        "tags": [
          "Articulo"
        ],
        "summary": "Obtiene artículos segun categoria",
        "description": "Obtiene articulos segun categoria",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id de categoria para mostrar",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "content": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/ArticuloResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    },
    "/articulo/imagenes/{id}": {
      "get": {
        "tags": [
          "Articulo"
        ],
        "summary": "Obtiene las imagenes de un Articulo",
        "description": "Obtiene las imagenes de un Articulo",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id de Articulo para mostrar las imagenes",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "content": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Articulo_ImagenResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "MarcaResponse": {
      "type": "object",
      "properties": {
        "idmarca": {
          "type": "integer",
          "example": 10
        },
        "descripcion": {
          "type": "string",
          "example": "Mi Marca"
        },
        "estado": {
          "type": "integer",
          "example": 1
        }
      }
    },
...
    "CategoriaIdResponse": {
      "type": "object",
      "properties": {
        "idcategoria": {
          "type": "integer",
          "example": 10
        },
        "descripcion": {
          "type": "string",
          "example": "Mi Categoria"
        },
        "estado": {
          "type": "integer",
          "example": 1
        },
        "idcategoriapadre": {
          "type": "integer",
          "example": 2
        }
      }
    },
...
    "Articulo": {
      "type": "object",
      "properties": {
        "idarticulo": {
          "type": "integer",
          "example": 10
        },
        "idmarca": {
          "type": "integer",
          "example": 10
        },
        "idtipoproducto": {
          "type": "integer",
          "example": 10
        },
        "idcategoria": {
          "type": "integer",
          "example": 10
        },
        "idunidadmedida": {
          "type": "integer",
          "example": 10
        },
        "nombre": {
          "type": "string",
          "example": "Mi Articulo"
        },
        "descripcion": {
          "type": "string",
          "example": "Mi Descripcion"
        },
        "et": {
          "type": "string",
          "example": "Mi et"
        },
        "precio": {
          "type": "number",
          "example": "99.99"
        },
        "imagen": {
          "type": "string",
          "example": "http:/www.images.com/img_01.jpg"
        },
        "stock": {
          "type": "integer",
          "example": 10
        },
        "sku": {
          "type": "string",
          "example": "Mi sku"
        },
        "estado": {
          "type": "integer",
          "example": 1
        }
      }
    },
    "ArticuloResponse": {
      "type": "object",
      "properties": {
        "idmarca": {
          "type": "integer",
          "example": 10
        },
        "idtipoproducto": {
          "type": "integer",
          "example": 10
        },
        "idcategoria": {
          "type": "integer",
          "example": 10
        },
        "idunidadmedida": {
          "type": "integer",
          "example": 10
        },
        "nombre": {
          "type": "string",
          "example": "Mi Articulo"
        },
        "descripcion": {
          "type": "string",
          "example": "Mi Descripcion"
        },
        "et": {
          "type": "string",
          "example": "Mi et"
        },
        "precio": {
          "type": "number",
          "example": "99.99"
        },
        "imagen": {
          "type": "string",
          "example": "http:/www.images.com/img_01.jpg"
        },
        "stock": {
          "type": "integer",
          "example": 10
        },
        "sku": {
          "type": "string",
          "example": "Mi sku"
        },
        "estado": {
          "type": "integer",
          "example": 1
        }
      }
    },
    "ArticuloIdResponse": {
      "type": "object",
      "properties": {
        "idmarca": {
          "type": "integer",
          "example": 10
        },
        "idtipoproducto": {
          "type": "integer",
          "example": 10
        },
        "idcategoria": {
          "type": "integer",
          "example": 10
        },
        "idunidadmedida": {
          "type": "integer",
          "example": 10
        },
        "nombre": {
          "type": "string",
          "example": "Mi Articulo"
        },
        "descripcion": {
          "type": "string",
          "example": "Mi Descripcion"
        },
        "et": {
          "type": "string",
          "example": "Mi et"
        },
        "precio": {
          "type": "number",
          "example": "99.99"
        },
        "imagen": {
          "type": "string",
          "example": "http:/www.images.com/img_01.jpg"
        },
        "stock": {
          "type": "integer",
          "example": 10
        },
        "sku": {
          "type": "string",
          "example": "Mi sku"
        },
        "estado": {
          "type": "integer",
          "example": 1
        },
        "tbl_articulo_imagen": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Articulo_ImagenResponse"
          }
        },
        "tbl_marca": {
          "$ref": "#/definitions/MarcaResponse"
        },
        "tbl_categoria": {
          "$ref": "#/definitions/CategoriaIdResponse"
        },
        "tbl_tipo_producto": {
          "$ref": "#/definitions/Tipo_ProductoResponse"
        },
        "tbl_unidad_medida": {
          "$ref": "#/definitions/Unidad_MedidaResponse"
        }
      }
    },
...
    "Articulo_ImagenResponse": {
      "type": "object",
      "properties": {
        "idarticuloimagen": {
          "type": "integer",
          "example": 10
        },
        "idarticulo": {
          "type": "integer",
          "example": 10
        },
        "imagen": {
          "type": "string",
          "example": "http:/www.images.com/img_01.jpg"
        }
      }
    },
...
    "Unidad_MedidaResponse": {
      "type": "object",
      "properties": {
        "idunidadmedida": {
          "type": "integer",
          "example": 10
        },
        "descripcion": {
          "type": "string",
          "example": "Mi Unidad de Medida"
        },
        "estado": {
          "type": "integer",
          "example": 1
        }
      }
    },
...
    "Tipo_ProductoResponse": {
      "type": "object",
      "properties": {
        "idtipoproducto": {
          "type": "integer",
          "example": 10
        },
        "descripcion": {
          "type": "string",
          "example": "Mi Tipo de Producto"
        },
        "estado": {
          "type": "integer",
          "example": 1
        }
      }
    },
...
    "InvalidResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "example": false
        },
        "error": {
          "type": "string",
          "example": "No se encontro registro"
        }
      }
    }
  }
}